requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
DateTime==5.5
Flask==3.0.3
fonttools==4.53.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jsonify==0.5
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
plotly==5.23.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
tenacity==9.0.0
typing==3.7.4.3
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
zope.interface==7.0.1
